refactor(amp): align ACP adapter spawn
