+++ packages/plugins/src/agents/impl/amp/index.ts
@@ -1,6 +1,3 @@
-import { mkdirSync, writeFileSync } from 'node:fs';
-import { tmpdir } from 'node:os';
-import { join } from 'node:path';
 import { definePlugin, registerPluginBehavior } from '@emdash/core/agents/plugins';
 import {
   ampMcpAdapter,
@@ -8,22 +5,14 @@ import {
   createFileDropPlugin,
   npmDependency,
 } from '@emdash/core/agents/plugins/helpers';
-import { createNativeAcpBehavior } from '../../helpers/acp-stdio';
-import { AMP_ACP_ADAPTER_SCRIPT } from './acp-adapter-script';
+import { connectStdioAcp } from '../../helpers/acp-stdio';
+import { resolveAmpAcpAdapterEntry } from './acp-adapter-entry';
 import { icon } from './icon';
 import { AMP_PLUGIN_CONTENT } from './plugin-file';
 
 const AMP_PLUGIN_PATH = '.amp/plugins/emdash-hook.ts';
-const AMP_ACP_ADAPTER_DIR = join(tmpdir(), 'emdash-amp-acp-adapter');
-const AMP_ACP_ADAPTER_PATH = join(AMP_ACP_ADAPTER_DIR, 'adapter.cjs');
 // Amp thread ids are prefixed with 'T-'; only accept those for resume.
 const validateSessionId = (id: string) => id.startsWith('T-');
-
-function ensureAmpAcpAdapterScript(): string {
-  mkdirSync(AMP_ACP_ADAPTER_DIR, { recursive: true });
-  writeFileSync(AMP_ACP_ADAPTER_PATH, AMP_ACP_ADAPTER_SCRIPT, 'utf8');
-  return AMP_ACP_ADAPTER_PATH;
-}
 
 export const plugin = definePlugin(
   {
@@ -92,11 +81,14 @@ export const provider = registerPluginBehavior(plugin, {
   // can create the remote thread and then leave the ACP turn stuck in `working`.
   // Shell out to Emdash's resolved Amp CLI directly until amp-acp prioritizes
   // AMP_CLI_PATH or Amp exposes a native ACP subcommand.
-  acp: createNativeAcpBehavior((ctx) => ({
-    command: process.execPath,
-    args: [ensureAmpAcpAdapterScript()],
-    env: { ELECTRON_RUN_AS_NODE: '1', AMP_CLI_PATH: ctx.cli },
-  })),
+  acp: {
+    buildSpawn: (ctx) => ({
+      command: process.execPath,
+      args: [resolveAmpAcpAdapterEntry()],
+      env: { ELECTRON_RUN_AS_NODE: '1', AMP_CLI_PATH: ctx.cli },
+    }),
+    connect: (io, toClient) => connectStdioAcp(io, toClient),
+  },
   prompt: {
     buildCommand: (ctx) =>
       buildStandardCommand(ctx, {
